fix(env-check): rename file
