+++ packages/environment-check/src/checks/get-installed.ts
